python/bot functions/requirements.txt
azure-functions
mysql-connector